requirements.txt
Django==2.2.8
django-phone-field
matplotlib
black
djangorestframework


